% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abonnement.R
\name{abonnement}
\alias{abonnement}
\title{Shiny module providing GUI and server logic for the subscription v2 tab}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing GUI and server logic for the subscription v2 tab
}
