% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plots}
\alias{plots}
\title{Shiny module providing GUI and server logic for the plot tab}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing GUI and server logic for the plot tab
}
