% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samlerapport.R
\name{samlerapport}
\alias{samlerapport}
\title{Shiny module providing GUI and server logic for the report tab}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing GUI and server logic for the report tab
}
