% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utsending.R
\name{utsending}
\alias{utsending}
\title{Shiny module providing GUI and server logic for the dispatch tab}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing GUI and server logic for the dispatch tab
}
