% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_veiledning.R
\name{veiledning}
\alias{veiledning}
\title{Shiny module providing GUI and server logic for the intro tab}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing GUI and server logic for the intro tab
}
